feat: implement CheckoutButton component for cart checkout functionality
