+++ app/(store)/cart/page.tsx
@@ -3,7 +3,8 @@
 import { api } from "@/trpc/react";
 import { useCart } from "@/hooks/cart-context";
 import CartList from "@/components/cart/list";
-import { Box, Button, Typography, Divider } from "@mui/material";
+import CheckoutButton from "@/components/cart/checkout-button";
+import { Box, Typography } from "@mui/material";
 
 export default function CartPage() {
 	const { cart } = useCart();
@@ -17,15 +18,6 @@ export default function CartPage() {
 			enabled: cart.length > 0, // prevents SSR error during build
 		},
 	);
-
-	let total = 0;
-	if (products) {
-		total = cart.reduce((sum, item) => {
-			const product = products.find((p) => p?._id === item.itemId);
-			if (!product) return sum;
-			return sum + product.price * (item.quantity ?? 1);
-		}, 0);
-	}
 
 	return (
 		<main>
@@ -57,31 +49,8 @@ export default function CartPage() {
 						</Typography>
 					</Box>
 				)}
-				{cart.length > 0 && (
-					<Box
-						sx={{
-							mt: 4,
-							display: "flex",
-							flexDirection: "column",
-						}}
-					>
-						<Divider sx={{ width: "100%", mb: 2 }} />
-						<Typography variant="subtitle1" sx={{ fontWeight: 600, mb: 2 }}>
-							Total: ${total.toFixed(2)}
-						</Typography>
-						<Button
-							variant="contained"
-							color="primary"
-							size="large"
-							disabled={cart.length === 0}
-							onClick={() => {
-								/* Checkout action placeholder */
-							}}
-							sx={{ minWidth: 180, fontWeight: 600 }}
-						>
-							Checkout
-						</Button>
-					</Box>
+				{cart.length > 0 && products && (
+					<CheckoutButton cart={cart} products={products} />
 				)}
 			</Box>
 		</main>
